Move logs related files to its own folder; update ownership paths
