+++ .github/CODEOWNERS
@@ -435,8 +435,8 @@ lerna.json @grafana/frontend-ops
 /public/testdata/ @grafana/grafana-frontend-platform
 /public/views/ @grafana/grafana-frontend-platform
 
-/public/app/features/explore/Logs.tsx @grafana/observability-logs
-/public/app/features/explore/LogsContainer.tsx @grafana/observability-logs
+/public/app/features/explore/Logs/Logs.tsx @grafana/observability-logs
+/public/app/features/explore/Logs/LogsContainer.tsx @grafana/observability-logs
 
 /public/app/features/explore/TraceView/ @grafana/observability-traces-and-profiling
