+++ public/app/features/explore/Logs/Logs.tsx
@@ -43,13 +43,13 @@ import { dedupLogRows, filterLogLevels } from 'app/core/logsModel';
 import store from 'app/core/store';
 import { ExploreId } from 'app/types/explore';
 
-import { LogRows } from '../logs/components/LogRows';
-import { LogRowContextModal } from '../logs/components/log-context/LogRowContextModal';
+import { LogRows } from '../../logs/components/LogRows';
+import { LogRowContextModal } from '../../logs/components/log-context/LogRowContextModal';
+import { SETTINGS_KEYS } from '../utils/logs';
 
 import { LogsMetaRow } from './LogsMetaRow';
 import LogsNavigation from './LogsNavigation';
 import { LogsVolumePanelList } from './LogsVolumePanelList';
-import { SETTINGS_KEYS } from './utils/logs';
 
 interface Props extends Themeable2 {
   width: number;
