+++ public/app/features/explore/Logs/LogsMetaRow.tsx
@@ -6,12 +6,11 @@ import { LogsDedupStrategy, LogsMetaItem, LogsMetaKind, LogRowModel, CoreApp, da
 import { reportInteraction } from '@grafana/runtime';
 import { Button, Dropdown, Menu, ToolbarButton, Tooltip, useStyles2 } from '@grafana/ui';
 
-import { downloadLogsModelAsTxt } from '../inspector/utils/download';
-import { LogLabels } from '../logs/components/LogLabels';
-import { MAX_CHARACTERS } from '../logs/components/LogRowMessage';
-import { logRowsToReadableJson } from '../logs/utils';
-
-import { MetaInfoText, MetaItemProps } from './MetaInfoText';
+import { downloadLogsModelAsTxt } from '../../inspector/utils/download';
+import { LogLabels } from '../../logs/components/LogLabels';
+import { MAX_CHARACTERS } from '../../logs/components/LogRowMessage';
+import { logRowsToReadableJson } from '../../logs/utils';
+import { MetaInfoText, MetaItemProps } from '../MetaInfoText';
 
 const getStyles = () => ({
   metaContainer: css`
