+++ public/app/features/explore/Logs/LogsVolumePanelList.tsx
@@ -14,11 +14,11 @@ import {
 } from '@grafana/data';
 import { Button, InlineField, useStyles2 } from '@grafana/ui';
 
-import { mergeLogsVolumeDataFrames, isLogsVolumeLimited, getLogsVolumeMaximumRange } from '../logs/utils';
+import { mergeLogsVolumeDataFrames, isLogsVolumeLimited, getLogsVolumeMaximumRange } from '../../logs/utils';
+import { SupplementaryResultError } from '../SupplementaryResultError';
+import { isTimeoutErrorResponse } from '../utils/logsVolumeResponse';
 
 import { LogsVolumePanel } from './LogsVolumePanel';
-import { SupplementaryResultError } from './SupplementaryResultError';
-import { isTimeoutErrorResponse } from './utils/logsVolumeResponse';
 
 type Props = {
   logsVolumeData: DataQueryResponse | undefined;
